ChatWithDoc/requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
pyPDF2
chromadb
faiss-cpu
langchain_google_genai
langchain